Move the main components into the containers folder given that they contain other components
